Remove AssimpProcessing, as this is purely internal and include it a protected methods in AssimpLoader
