+++ source/gloperate-assimp/include/gloperate-assimp/AssimpLoader.h
@@ -7,6 +7,10 @@
 #include <gloperate-assimp/gloperate-assimp_api.h>
 
 
+struct aiMesh;
+struct aiScene;
+
+
 namespace gloperate_assimp
 {
 
@@ -38,6 +42,34 @@ public:
     virtual std::vector<std::string> loadingTypes() const override;
     virtual std::string allLoadingTypes() const override;
     virtual PolygonalGeometry * load(const std::string & filename, std::function<void(int, int)> progress) const override;
+
+
+protected:
+    /**
+    *  @brief
+    *    Convert all meshes from an ASSIMP scene into gloperate meshes
+    *
+    *  @param[in] scene
+    *    ASSIMP scene (must be valid!)
+    *  @param[in] numMeshes
+    *    Maximum number of meshes that are imported (if 0, all found meshes are imported)
+    *
+    *  @return
+    *    List of meshes, which must be destroyed by the caller
+    */
+    std::vector<PolygonalGeometry *> convertGeometries(const aiScene * scene, size_t numMeshes = 0) const;
+
+    /**
+    *  @brief
+    *    Convert ASSIMP mesh into gloperate mesh
+    *
+    *  @param[in] mesh
+    *    ASSIMP mesh (must be valid!)
+    *
+    *  @return
+    *    Mesh, must be destroyed by the caller
+    */
+    PolygonalGeometry * convertGeometry(const aiMesh * mesh) const;
 };
